renaming editors
