+++ Assets/Mirage/Editor/Logging/LogSettingsSOEditor.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 85ad7b27a4ecdb345a945d860844eaf1
+guid: bd876bb5841c79e4fbeefba3be177f63
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
